Port AOM commits: 1. 6e61c8e4b 2. d2d6931c3

+++ Source/Lib/Encoder/ASM_AVX512/pickrst_avx512.c
@@ -3358,7 +3358,7 @@ void eb_av1_compute_stats_highbd_avx512(int32_t wiener_win, const uint8_t *dgd8,
     eb_aom_free(d);
 }
 
-static INLINE __m512i pair_set_epi16_avx512(uint16_t a, uint16_t b) {
+static INLINE __m512i pair_set_epi16_avx512(int32_t a, int32_t b) {
     return _mm512_set1_epi32((int32_t)(((uint16_t)(a)) | (((uint32_t)(b)) << 16)));
 }
